+++ Dockerfile
@@ -1,1 +1,1 @@
-FROM tas-tools-ext-01.ccr.xdmod.org/xdmod-9.0.0:centos7.8-0.1
+FROM tas-tools-ext-01.ccr.xdmod.org/xdmod-9.5.0:centos7.9-1.0
