+++ bin/xdmod-upgrade
@@ -26,7 +26,7 @@ ini_set('memory_limit', -1);
  * @var array
  */
 $supportedUpgrades = array(
-    '9.0.0' => '9.5.0'
+    '9.5.0' => '10.0.0'
 );
 
 /**
