+++ classes/OpenXdmod/Migration/Version950To1000/DatabaseMigration.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Update database from version 9.5.0 To 10.0.0.
+ */
+
+namespace OpenXdmod\Migration\Version950To1000;
+
+use OpenXdmod\Migration\DatabasesMigration as AbstractDatabasesMigration;
+use OpenXdmod\Setup\Console;
+
+/**
+ * Migrate databases from version 9.5.0 to 10.0.0.
+ */
+class DatabasesMigration extends AbstractDatabasesMigration
+{
+    public function execute()
+    {
+        parent::execute();
+    }
+}
