+++ configuration/portal_settings.ini
@@ -6,7 +6,7 @@ contact_page_recipient = ""
 tech_support_recipient = ""
 
 ; The version number is updated during the upgrade process.
-version = "9.5.0"
+version = "10.0.0"
 
 debug_mode = "off"
 debug_recipient = ""
