+++ docs/_config.yml
@@ -36,8 +36,8 @@ defaults:
       type: "pages"
     values:
       layout: "page"
-      prev_sw_version: "9.0.0"
-      sw_version: "9.5.0"
+      prev_sw_version: "9.5.0"
+      sw_version: "10.0.0"
       style: "effervescence"
       tocversion: "toc"
