+++ open_xdmod/modules/xdmod/build.json
@@ -1,6 +1,6 @@
 {
     "name": "xdmod",
-    "version": "9.5.0",
+    "version": "10.0.0",
     "release": "1.0",
     "files": {
         "include_paths": [
