+++ tests/ui/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "automated-tests",
+  "private": true,
+  "name": "ui-tests",
   "version": "1.0.0",
-  "description": "Selenium Tests for XDMod",
-  "main": "index.js",
+  "description": "UI Tests for Open XDMoD",
   "engines": {
     "node": ">=6.0"
   },
@@ -15,7 +15,7 @@
   "dependencies": {
     "chai": "4.1.2",
     "cheerio": "1.0.0-rc.3",
-    "chromedriver": "84.0.1",
+    "chromedriver": "91.0.0",
     "mocha": "5.2",
     "require-dir": "1.0",
     "selenium-standalone": "6.16.0",
